third_party/blink/web_tests/external/wpt/css/css-masking/parsing/mask-valid.sub-expected.txt

This is a testharness.js-based test.
Found 10 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] e.style['mask'] = "none" should set mask-border-outset
  assert_equals: mask-border-outset should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-repeat
  assert_equals: mask-border-repeat should be canonical expected (string) "stretch" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-slice
  assert_equals: mask-border-slice should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-source
  assert_equals: mask-border-source should be canonical expected (string) "none" but got (undefined) undefined
[FAIL] e.style['mask'] = "none" should set mask-border-width
  assert_equals: mask-border-width should be canonical expected (string) "auto" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-outset
  assert_equals: mask-border-outset should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-repeat
  assert_equals: mask-border-repeat should be canonical expected (string) "stretch" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-slice
  assert_equals: mask-border-slice should be canonical expected (string) "0" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-source
  assert_equals: mask-border-source should be canonical expected (string) "none" but got (undefined) undefined
[FAIL] e.style['mask'] = "none, linear-gradient(to left bottom, red, blue) padding-box" should set mask-border-width
  assert_equals: mask-border-width should be canonical expected (string) "auto" but got (undefined) undefined
Harness: the test ran to completion.
